Move settings related classes inside prefs package
